refactor:(#106) 변수명, 파일명 오타 및 의미에 맞게 수정
