+++ frontend/src/page/detailIssuePage/detailHeader/DetailIssueHeader.tsx
@@ -2,7 +2,7 @@ import React, { useEffect } from 'react';
 import styled from 'styled-components';
 import { useRecoilValue, useRecoilState } from 'recoil';
 import { DetailIssueType } from 'page/detailIssuePage/DetailIssuePage';
-import { timeChecker } from '../../../util/util';
+import { timeChecker } from '../../../util/timeUtil';
 import DetailIssueStatus from './DetailIssueStatus';
 import HeaderViewMode from 'page/detailIssuePage/detailHeader/HeaderViewMode';
 import HeaderEditMode from 'page/detailIssuePage/detailHeader/HeaderEditMode';
