+++ frontend/src/util/timeUtil.ts
